rc1
git-svn-id: https://svn.codehaus.org/enunciate/trunk/enunciate@683 c2b6a248-b21b-0410-9d80-82131e002806

+++ core/src/main/java/org/codehaus/enunciate/main/Enunciate.java
@@ -876,18 +876,20 @@ public class Enunciate {
    * @param to   The destination directory.
    */
   public void copyDir(File from, File to) throws IOException {
-    File[] files = from.listFiles();
+    if (from != null && from.exists()) {
+      File[] files = from.listFiles();
 
-    if (!to.exists()) {
-      to.mkdirs();
-    }
-
-    for (File file : files) {
-      if (file.isDirectory()) {
-        copyDir(file, new File(to, file.getName()));
+      if (!to.exists()) {
+        to.mkdirs();
       }
-      else {
-        copyFile(file, new File(to, file.getName()));
+
+      for (File file : files) {
+        if (file.isDirectory()) {
+          copyDir(file, new File(to, file.getName()));
+        }
+        else {
+          copyFile(file, new File(to, file.getName()));
+        }
       }
     }
   }
